Backend/requirements.txt
flask>=3.0
gunicorn>=21.0
flask-cors>=4.0
pymongo>=4.0
python-dotenv>=1.0
certifi>=2024.0
